projects/zillow-tracker/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
lxml==4.9.3
python-dotenv==1.0.0
schedule==1.2.0